+++ packages/audit/src/interfaces.ts
@@ -1,3 +1,18 @@
+export interface AuditBody {
+  // Who
+  actorRef: string;
+  orgRef: string;
+
+  // Where (it can be an array since the request can jump between nodes, and http keeps record of their IPs)
+  ip: string | string[];
+
+  // What
+  action: string;
+  succeed: boolean;
+  targetRef?: string;
+  metadata?: Metadata;
+}
+
 export enum CommonMetadataKeys {
   EMPLOYEE_ID = "employee_id",
   EMPLOYEE_IDS = "employee_ids",
@@ -16,25 +31,3 @@ export const REQUIRED_FIELDS = [
   "action",
   "succeed",
 ];
-
-export interface AuditBody {
-  // Who
-  actorRef: string;
-  orgRef: string;
-
-  // Where (it can be an array since the request can jump between nodes, and http keeps record of their IPs)
-  ip: string | string[];
-
-  // What
-  action: string;
-  succeed: boolean;
-  targetRef?: string;
-  metadata?: Metadata;
-}
-
-// When will be set by logger (log time)
-// Where will be set by logger (host name)
-
-// Articles that we based to create this:
-// https://workos.com/blog/the-developers-guide-to-audit-logs-siem
-// https://www.strongdm.com/blog/audit-log-review-management