+++ packages/core/src/app/appIdentifier.ts
@@ -1,0 +1,5 @@
+export function appIdentifier() {
+  const appName = process.env.API_ORIGIN_NAME || "undefined";
+  const env = process.env.NODE_ENV || "development";
+  return `${appName}:${env}`;
+}