+++ packages/logger/src/index.ts
@@ -1,5 +1,5 @@
 export { createLogger } from "./factory";
 export * from "./config";
 export * from "./logger";
-export * from "./interfaces";
 export * from "./mock";
+export * from "./interfaces";