+++ packages/logger/src/interfaces.ts
@@ -1,9 +1,14 @@
-export enum LogLevel {
-  TRACE = "trace",
-  DEBUG = "debug",
-  INFO = "info",
-  WARN = "warn",
-  ERROR = "error",
+export interface ILogger {
+  trace(code: string, context?: Context): void;
+  debug(code: string, context?: Context): void;
+  info(code: string, context?: Context): void;
+  warn(code: string, context?: Context): void;
+  error(code: string, context?: Context): void;
+  child(context?: Context): ILogger;
+}
+
+export interface Context {
+  [key: string]: any;
 }
 
 export enum LogType {
@@ -12,15 +17,10 @@ export enum LogType {
   AUDIT = "audit",
 }
 
-export interface Context {
-  [key: string]: any;
-}
-
-export interface ILogger {
-  trace(code: string, context?: Context): void;
-  debug(code: string, context?: Context): void;
-  info(code: string, context?: Context): void;
-  warn(code: string, context?: Context): void;
-  error(code: string, context?: Context): void;
-  child(context?: Context): ILogger;
+export enum LogLevel {
+  TRACE = "trace",
+  DEBUG = "debug",
+  INFO = "info",
+  WARN = "warn",
+  ERROR = "error",
 }