+++ packages/serverless/src/handlers/wrappers/withContext.ts
@@ -1,0 +1,34 @@
+import { ILogger } from "@alanszp/logger";
+import { Audit } from "@alanszp/audit";
+import { SharedContext } from "@alanszp/shared-context";
+import { Context } from "aws-lambda";
+import middy, { MiddlewareObj } from "@middy/core";
+import cuid from "cuid";
+import { appIdentifier } from "@alanszp/core";
+import { Lambda } from "../../types";
+
+export function withContext<TEvent, TContext extends Context, TResult>(
+  sharedContext: SharedContext,
+  handler: Lambda<TEvent, TContext, TResult>,
+  middlewares: MiddlewareObj<TEvent, TResult>[],
+  baseLogger: ILogger,
+  audit: Audit
+): Lambda<TEvent, TContext, TResult> {
+  return (event, context) =>
+    sharedContext.run(
+      async () => {
+        const handlerWithExecutionContext = middlewares.reduce(
+          (currentHandler, newMiddleware) => currentHandler.use(newMiddleware),
+          middy(handler)
+        );
+        return handlerWithExecutionContext(event, context);
+      },
+      {
+        logger: baseLogger,
+        audit: audit.withState(),
+        lifecycleId: cuid(),
+        lifecycleChain: appIdentifier(),
+        contextId: cuid(),
+      }
+    );
+}