+++ packages/serverless/src/types/lambda.ts
@@ -1,0 +1,6 @@
+import { LambdaContext } from "./middyTypes";
+
+export type Lambda<TEvent, TContext extends LambdaContext, TResult> = (
+  event: TEvent,
+  context: TContext
+) => Promise<TResult> | void;