+++ packages/shared-context/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./sharedContext";
+export * from "./SharedContextNotInitializedException";